WIP --nothing working. Record keeping only
